Changes to Assignment/AssignmentManager
Refined overloaded operators, added an isComplete function to print a
completed assignment, and a conversion function for a Date into a string
in the Assignment class.  Made a checker for duplicate assignments in
the addAssignment function. Created a separate namespace for the enum
type and its conversion-from-a-string function. Still need work on a
little exception handling and objectifying code in the switch cases
under UserInterface.cpp.