Add function edit
If the assignment being added was the biggest one and needed to go onto
the end of the list, it would not get added. The for loop would exit
because the iter was at the_list.end()

Fixed by adding in a bool statement that switches to true after the
assignment has been added to the list. If it stays false the assignment
will be added to the end of the list via the_list.push_back(assingmnet);